requirements-dev.txt
-r requirements.txt
pytest==2.6.0
pytest-cov==1.7.0
pytest-django==2.6.2
Werkzeug==0.9.6
ptpython==0.1
django-extensions==1.4.8
